GEN_AI/requirements.txt
streamlit 
google-generativeai 
python-dotenv
